move signature changes to xmtp_cryptography
