+++ xmtp/Cargo.toml
@@ -11,7 +11,7 @@ serde_json = "1.0.96"
 thiserror = "1.0.40"
 vodozemac = {git= "https://github.com/xmtp/vodozemac", branch="dev"}
 xmtp_proto = { path = "../xmtp_proto", features = ["xmtp-v3-message_contents"] }
-xmtp_crypto = { path = "../xmtp_crypto"}
+xmtp_cryptography = { path = "../xmtp_cryptography"}
 hex = "0.4.3"
 
 [dev-dependencies]
