+++ xmtp_crypto/src/lib.rs
@@ -4,8 +4,6 @@ pub mod k256_helper;
 pub mod signature;
 pub mod traits;
 
-pub mod utils;
-
 #[cfg(test)]
 mod tests {
     use crate::*;
